requirements.txt
Pillow==9.4.0
requests==2.25.1
streamlit==1.19.0
